Added the new network interface, now working copydata through the caches.
Each remote node is now represeted by a thread, using a ClusterRemoteNode PE
and a ClusterThread. CopyData info on workdescriptors is well handled by
the runtime using the cache and correctly performs the required memory transfers
among nodes. Code has been added to the cache in order to handle the case
where different caches have to keep the same entry coherent.
Network operations are done now through the NetworkAPI interface. gasnetapi.[c,hpp]
contain the gasnet implementation of the API.
[BUG?] _untieMaster = false when initializing the system is required in order to
keep the master WD running on the first thread, otherwise it can occur that a
ClusterThread tries to switchTo() the master WD, which is an unwanted situation.