GASNet segments, start/stop tracing.
GASNet requires that big messages are received on a given address
range, so now ClusterDevice::allocate does not use malloc, instead
it selects a memory chunk in the given range (it is mmaped during
gasnet_init()).
The call to gasnet_finalize through _net->finalize() has been moved
further since it terminates the process, this was preventing traces
from being generated.