New CachePolicy implemented: AsynchronousWriteThroughPolicy.
It takes into account that copies may be delayed and can be executed a while after unregisterCacheAccess is called. It also knows that data on main memory may not be the most updated version and when this case is detected, it waits for the data to be updated properly (not the optimal solution, but at least seems to work :-) ).