Using SimpleAllocator class to manage GPU memory.
The whole GPU memory is allocated at once, and is managed through Simple Allocator class.
It may happen that memory space is partitioned in a way that a block of data could fit if
free space was contiguous, but we find that the free space is partitioned. In this case,
the application will fail: it must be solved.
Currently, the GPU memory is not freed at the end of the application. This must be solved,
too.