When GPU's are busy, avoid causing a segmentation fault.
In fact, this specific error is controlled when CUDA streams are created
(if we need to create them) or when we malloc the whole GPU memory. By
this way, when the 'cudaErrorDevicesUnavailable' is detected, the runtime
aborts immediately the execution.
If streams cannot be created due to any other reason, we just print a
warning message, and let the application go on.

Moreover, input and output memory transfer lists are initialized to the
default memory transfer list (which does nothing) instead of being
initialized to NULL. This was causing the segmentation fault.