Disable GPUs when execute test_single.cpp.
This test creates SMP tasks and calls a team barrier inside these tasks,
assuming that all threads are SMP threads and will be able to run the tasks.
If GPUs are enabled, GPU threads will never execute those tasks and will
never reach the barrier, so the test will hang.