nanoxrun.sh, transfer of traces, prefetch code
Added the script nanoxrun.sh to help with the execution of applications in interactive and batch mode.
Trace files are transfered from remote nodes to the master node. This helps generating a whole trace of an execution.
Prefetch code has been added to send the data of a task to a node immediatly after a previous task has been executed.
However the code is currently commented.