Reset CUDA errors that may have occurred inside the runtime initialization.
Reset the errors so that the user application cannot see them if it
calls cudaGetLastError().

The user application still will be able to see the errors occured
during its execution, but at least we avoid the ones occured during
initialization.

We do not have to worry about these errors, because returning values
from each call to CUDA are checked and treated properly.