Split WD::finish() into 2 phases, as some things should be done before WD::done() is called.
WD::done() makes the copy-out's of the task, signals its parent and after
exiting this function, the WD can be potentially removed.
However, in order to register the WD's execution time and process it, we
need to get the final execution time after WD's copies have been performed,
but we don't want the WD be removed at that time.
So, it is better to split the function into two parts:
- WD::finish() --> wait for its children, copy out data and get the
  execution time.
- WD::done() --> notify the parent and the WG.