Support direct copies between GPUs. (Fixes #530)
As CUDA 4.0 allows direct copies between GPU devices without the need
to copy the data to host, added the code to enable this new feature
in the runtime. THe runtime still supports CUDA 3.2 as well.

There is a new function in the device that performs this kind of
copies: copyDevToDev().

In the cache, the invalidate() function has been split into two
different functions:
- invalidateAndFlush() does the same as the old invalidate() function
  did (invalidate data and copy back to host)
- invalidate() function only invalidates data, but does not copy
  them back to host