New Directory, bug fixes
New Directory initial development phase, not doing anything usefull but working ok so far,
some cases need to be added to memorymap in order to merge all possible cases.

WorkGroup seed moved to System class to ensure that is not double initialized by being global.

Remote Workgroup creation changed to ensure that "mainWD" is the first WD of the system.

Header clean up to use fwd, or decl if possible.