#581: The scheduler can be halted.
First implementation of the feature. It works when:
· The scheduler is halted (with no tasks running), some tasks are submitted and then the scheduler is started again.
· Some tasks are submitted and while these are still running, the scheduler is halted, then fed some more tasks, and then started again.

Two new tests created in order to check this. They work with both the cilk scheduler policy and the rest of the schedulers of the test generator.

The priority test has now a lock in the critical section.
A new priority test has been created that relies on scheduler start/stop functionality.