Set the CUBLAS execution stream for each GPU thread. (Closes #564)
But still the programmer needs to call cublasSetStream() function just
before any call to the CUBLAS library. There is no way for Nanox to
do it: calling cublasSetStream() just before a GPU thread calls WD's
outline has no effect.

Plus, if cublasSetStream() is not called by the programmer, calls
to CUBLAS library are not set to the appropriate stream and this makes
Nanox to incorrectly detect the end of the task: Nanox synchronizes
with its execution stream, but the CUBLAS call is sent to the default
stream (NULL stream), so calling cudaStreamSynchronize() returns
immediately. Traces reflect this behavior.