Changes in WD's _activeDevice.
It is incorrect to choose the device depending on its DeviceData, as
it might be more than one implementation per device: we need something
more specific to distinguish implementations. An index is used instead.
The index matches the position of the implementation in the WD's _devices
array.

Another member has been added to WD class: _activeDeviceIdx:
- If there is an active device, its value is the index of the active
  device in _devices array.
- If there is no active device, its value is the total number of
  implementations for the given WD (that is _devices array size).