Performance improvement.
When atPrefetch() is called, if no immediate successor is found, call
atIdle() instead of returning NULL directly. Maybe atIdle() finds
a ready task that the thread will be able to run.

This change is noticeable when combined with overlap of data transfers
and computation, since it is very important to get a task from
atPrefetch() in order to actually overlap data transfers with
computation.

If NULL was returned, the thread would have called atIdle() once the
current task was finished and it would have probably got the same task
that atPrefetch() is returning with this commit, but since the previous
task was already finished, there was no chance to overlap transfers and
computation.