Avoid Nanox+GPU applications crash at the end of their execution. (Fixes #678)
A new boolean member has been added to disable MemTracker mechanism and avoid
false and confusing 'Trying to free invalid pointer' messages.

MemTracker is disabled at the end of the execution, just right after memory
statistics (if enabled) are printed out. This way, MemTracker still works for
pointers that it has allocated, but does not fail when it tries to free a
pointer that it has not registered (because, actually, pointer's memory has been
allocated before MemTracker's creation).