Add a flag to wd_props to zero the chunk
Add a new field clear_chunk to wd_props that selects if the chunk
must be memset to zero or not.

This is needed by Mercurium since in Fortran, ALLOCATABLE fields may be
emitted during the OpenMP/OmpSs transformation. Such fields must be
initialized to zero when they are not allocated yet.

This commit fixes #715