Avoid segmentation fault when AsyncThread has no team. (Fixes #797)
Check at thread's runDependent() method whether the thread belongs to a team
or not. If not, the thread must exit because threads inheriting from
AsyncThread need a team to get tasks.

Consequently, AsyncThread::runDependent() method is not needed by now, so it
has been removed.

Note: this commit needs #798 to work properly.