Make WDDeque class aware of its WD's device types.
The class has been extended with a counter for each device in the list _dq.
This information prevents threads from visiting all the elements in the list
if there are no compatible WDs: From now on, these threads will only check
the value of an atomic member instead of acquiring a lock to go through all
list elements.

This is the default behavior, but can be switched back to the old one if
object's constructor is initialized with "false" value.