Invalidation bug
When a region was initialized on a cache, and then a subregion of it
was sent to anoher cache, an invalidation of the first region was
not properly handled. Only the non replicated part was being copied
to the host address space, but metadata was updated indicating that
the whole region was actually there. When a task was requesting
the subchunk, a copy was being generated from the host, which was
invalid because it wasn't there.