Cluster thread does not call to waitInputs anymore
waitInputs blocks and is not suitable for the cluster
thread which is asynchronous. A new call testInputs checks
if the data is ready without blocking and the thread manages
the set of workdescriptors that are waiting for data to
be executed. The previous behavior could cause deadlocks
when workdescriptors waited for copies issued by other
threads (when using "affinity support").