Fixing compilation problems in src/utils Makefile's.
The variable @USED_VERSION@ is not set properly for each library
version (performance, debug, instrumentation, instrumentation-debug).
This causes compilation errors. For example, the performance version of
libnanox is used to compile src/utils/nanox.cpp in debug version.

To fix this error, paths to libraries are now hard-coded in Makefile's.

This problem only shows up in src/utils directory, because it is the
only directory where @USED_VERSION@ variable is used in Makefile's. In
the rest of directories, the name of the version is already hard-coded.