Make affinity-ready scheduler aware of the architectures each memory space supports.
With this improvement, if a WD can only be run in one of the memory spaces of the
system, it will be added to that memory space's queue and affinity will not be
taken into account because it would not make sense in this case.

In addition, WDs will only be pushed to compatible queues (where there is at least
one PE that will be able to run the WD). So, affinity and compatibility are now
taken into account to avoid predictable (and unnecessary) work stealing.