Changes for gpu with presend, bug fix.
Fixed a race condition when registering a new region id.
Changes to support the prefetching path of the gpu architecture.
WDDeque now has a method to execute a function to all its WDs,
this is needed to call testInputs to prefetched WDs and avoid
deadlocks (WDs waiting for operations initiated by prefetched WDs).