Enhance GPU-side instrumentation.
CUDA 5.0 events and callbacks are used to better instrument what happens
on the GPU side.

More streams are created when instrumentation is enabled so that the
instrumentation code does not interfere with application's execution and
does not delay kernel executions.

By now, only copies (both input and output) are instrumented using this
method.