New version of the graph plugin
In this version, all the empty subgraphs generated for convenience
are not inserted into the final dot graph.
Also, consecutive subgraphs which are separated with a taskwait
are connected vertically to keep the execution order.

This commit also changes the name of the graph. So far it always was
"graph.dot" but now it will be "graph_BINARY_NAME.dot".