Enable CUBLAS initialization automatically from Mercurium. (Fixes #892)
From now on, CUBLAS will be initialized if it is set manually with the
usual configuration option (--gpu-cublas-init) / environment variable
(NX_GPUCUBLASINIT) OR if Mercurium defines the weak symbol 'gpu_cublas_init'.

It is an OR condition, so if 'gpu_cublas_init' symbol is defined, the user
will not be able to disable CUBLAS initialization through the configuration
option or environment variable.