First new graph plugin implementation
To be able to print the dependency graph in different fancy ways,
this commit implements a new version of the plugin that stores
the structure of the graph in memory during the execution and
then prints it to a dot format.
This way we can print different versions of the graph.

The option to enable this implementation is --instrumentation=new-graph.