Fix OMP_NUM_THREADS for OmpSs applications.
Using OMP_NUM_THREADS for OmpSs applications is discouraged,
two flags can be used to set the number of SMP workers on an
OmpSs application:
--smp-cores <n>: specifies how many hardware cores can access.
If no other arguments are specified, Nanos++ will create one worker
thread per SMP core that does not have a support thread created
(support threads are, gpu threads, cluster threads, etc).
The default value is the number of cores that are present
on the system. This is usually provided by the OS.

--smp-workers <n>: specifies the number of worker threads that the
runtime has to create. It is expected that this number is less than
or equal to the number of smp cores specified. If no value is specified,
as many worker threads as possible will be created, which it will
be the number of smp cores minus the support threads created by Nanos++.