** Temporal big ** change in GPU instrumentation.
The events emitted by CUDA callbacks are registered in another Extrae buffer,
so we avoid the race conditions of having 2 different threads writing to the
same Extrae buffer.

This means that the GPU events have been split into 2 different thread lines
in Paraver.

THIS IS VERY EXPERIMENTAL, as the functions called in order to enable this
behavior are very experimental and may crash in some situations.