Enqueue (CUDA) dependent tasks to the same CUDA execution stream.
This last dependency checking (needed for tasks returned from getImmediateSuccessor())
was done at the AsyncThread level, but in fact, in CUDA, this control can be left to
the stream mechanism, so we reduce a little bit the runtime overhead for immediate
successor tasks.

The WD has a new member to store the information related to the stream where it has
been sent.