Plugins must report the number of SMPThreads they will use
When they reserve a SMPProcessor they must call setNumFutureThreads
so the SMP Plugin knows the ammount of threads that will be created.
This is needed because the instrumentation plugin needs this value
to initialize correctly. However, no threads can be created before
initializing the instrumentation plugin, so this information must
be given in advance.