Solved problem with CUDA tests in Jenkins.
As the default environment uses CUDA 4.1, but actually OmpSs is compiled
with CUDA 5.0, there was a missmatch when tests were executed: CUDA tests
were compiled with CUDA 4.1, but the runtime was using CUDA 5.0. This
caused a segmentation fault at the very beginning of the test execution.

The solution is to use the whole CUDA path (set in configure process with
--with-cuda) to invoke nvcc to compile CUDA tests.