Changing the tpye of an internal data structure used by task reductions
I have been testing the performance of some benchmarks changing some internal
data structures. The best performance was obtained using an std::vector. This
commit changes the original list to a vector and adapts the code to this new
data structure.

Note that in the new version we are traversing the vector in a reversed order
(using reverse_iterators). The reason to do it in this way is that we are introducing
the new reductions at the end of the vector using the push_back method.


See #941