This commit define a new API call: nanos_exec_profile_kernel
This API call performs a profiling operation of a given kernel to find out what are the best configuration in terms of work-group parameters.
In order to do the profiling, we launch several executions of the same kernel according to the passed limits.

So far, we are manually providing those limits for unidimensional arrays: values between 1 and 2.