Fix task parameter MPI_Type_free on mpi_finalize
There are some cases in which some tasks are not executed
in a given process. This can produce that not all the entries
in the datatype cache are being used, thus a NULL pointer is
left. There were no checks for NULL pointers in datatype free
at nanosMPIFinalize, so MPI was complainting about this.