Fixed several errors and minor bugs
CacheCommands and regular Commands use a different tag on
their main channel to avoid collisions.
SingleDispatcher now takes a tag as template parameter so that
we can specify the different tags we created for each command
payload.
Deleted virtual constructors from request and persistent_request
to make them standard layout types.
Substituted Init and Finish commands in the remaining parts of the
code.