Implemented C++ wrappers for mutex and cond vars
This includes Mutex, LockGuards and UniqueLock that are used to
warrantee mutual exclusion access in critical sections and
ConditionVariable for thread synchronization.
They are implemented using pthread API and trying to be as close
as possible to the implementations provided by C++11 standard.
FileLock is a mutex implemented with files that was already being
used in mpi arch for interprocess mutual exclusion.