Print 'disabled' if feature is not available
When configure checks are finished, a summary is printed to inform
the user which features have been enabled or not, depending on
configuration flags, check results, etc.
If a feature was not present, it was not specified whether the
feature was present or not (just empty text).
Therefore, we added 'disabled' and 'enabled' words when a feature
such as allocator, memtracker, etc. is present.