TaskEnd control message is now received asynchronously
The receive is issued before the init control message is sent,
so that we make sure the reception is not produced before (eager
mode).
This way, it is possible to change MPI_Iprobe with an MPI_Waitsome
or MPI_Testsome for checking if a task has finished (checkTaskEnd).
This routines are way more efficient than MPI_Iprobe and we can
process task finalization in batchs.