Simplified MPIThread
Since RemoteSpawn contains all the necessary common information
of threads that share the same devices, it is no longer necessary
to keep most of the data structures that provided synchronization,
as well as the list of peer threads and processors.
The latter are contained in RemoteSpawn class, so they are still
accessible and coherent to all the threads in a group.