Use MPI_COMM_WORLD in MPI dev to dev transfers
In MPI, Intecommunicators are only used to communicate between
process in both ends (e.g. master to slave or slave to master).
If the application has to communicate between processes of the same
end (e.g. master to master or slave to slave), MPI_COMM_WORLD or
an equivalent regular communicator has to be used instead.